contrib guide: move pull-requests.md from devel/ to guide/
Understanding the pull request process is central to being a successful
contributor.  This commit moves it from the more technical k8s internals
detailed devel/ area over to the guide/ where it will get more
visibility and be better curated by virtue of that.

This patch just moves the file and leaves a tombstone reminder
behind. Subsequent patches are needed to scrub the content for accuracy,
readability and completeness.