contrib guide: add release notes document
If a contributor reads the pull request template upon creating a PR then
they will get hints, but ahead of a PR there has not been sufficient
documentation of when release notes are required for a change, how a
contributor should indicate that a change needs notes, or what the note
text should be.  To help address this, this commit adds a first set of
release notes documentation for reference by the contributor guide.